Removes useJSONAPI for compatibility with custom formatters
