+++ addon/factory-guy.js
@@ -81,6 +81,11 @@ var FactoryGuy =  Ember.Object.extend({
       return factory.get('fixtureBuilder');
     }
   }),
+  updateHTTPMethod: Ember.computed('fixtureBuilder', {
+    get() {
+      return this.getWithDefault('fixtureBuilder.updateHTTPMethod', 'PUT');
+    }
+  }),
   /**
    ```javascript
