+++ addon/jsonapi-fixture-builder.js
@@ -1,3 +1,4 @@
+import Ember from 'ember';
 import FixtureBuilder from './fixture-builder';
 import JSONAPIFixtureConverter from './jsonapi-fixture-converter';
 import JSONAPIAttributeTransformer from './jsonapi-attribute-transformer';
@@ -7,6 +8,12 @@ import JSONAPIAttributeTransformer from './jsonapi-attribute-transformer';
  */
 var JSONAPIJsonBuilder = function (store) {
   FixtureBuilder.call(this, store);
+
+  this.updateHTTPMethod = 'PATCH';
+
+  this.extractId = function (modelName, payload) {
+    return Ember.get(payload, 'data.id');
+  };
 
   this.convertForBuild = function (modelName, fixture) {
     var convertedFixture = new JSONAPIFixtureConverter(store).convert(modelName, fixture);
