+++ addon/mock-update-request.js
@@ -1,21 +1,6 @@
 import $ from 'jquery';
 import FactoryGuy from './factory-guy';
 
-/**
- Map single object to response json.
-
- Allows custom serializing mappings and meta data to be added to requests.
-
- @param {String} modelName model name
- @param {Object} json Json object from record.toJSON
- @return {Object} responseJson
- */
-function mapFind(modelName, json) {
-  var responseJson = {};
-  responseJson[modelName] = json;
-  return responseJson;
-}
-
 var MockUpdateRequest = function(url, model, options) {
   var status = options.status || 200;
   var succeed = true;
@@ -53,15 +38,14 @@ var MockUpdateRequest = function(url, model, options) {
     } else {
       // need to use serialize instead of toJSON to handle polymorphic belongsTo
       var json = model.serialize({includeId: true});
-      this.responseText = FactoryGuy.get('fixtureBuilderFactory').useJSONAPI() ? json : mapFind(model.constructor.modelName, json);
+      this.responseText = FactoryGuy.get('fixtureBuilder').normalize(model.constructor.modelName, json);
       this.status = 200;
     }
   };
-  var requestType = FactoryGuy.get('fixtureBuilderFactory').useJSONAPI() ? 'PATCH' : 'PUT';
   var requestConfig = {
     url: url,
     dataType: 'json',
-    type: requestType,
+    type: FactoryGuy.get('updateHTTPMethod'),
     response: this.handler
   };
