+++ addon/rest-fixture-builder.js
@@ -1,3 +1,4 @@
+import Ember from 'ember';
 import FixtureBuilder from './fixture-builder';
 import RESTFixtureConverter from './rest-fixture-converter';
 /**
@@ -7,6 +8,25 @@ import RESTFixtureConverter from './rest-fixture-converter';
  */
 var RESTFixtureBuilder = function(store) {
   FixtureBuilder.call(this, store);
+  /**
+   Map single object to response json.
+
+   Allows custom serializing mappings and meta data to be added to requests.
+
+   @param {String} modelName model name
+   @param {Object} json Json object from record.toJSON
+   @return {Object} responseJson
+   */
+  this.normalize = function(modelName, payload) {
+    var responseJson = {};
+    responseJson[modelName] = payload;
+    return responseJson;
+  };
+
+  this.extractId = function(modelName, payload) {
+    return Ember.get(payload, `${modelName}.id`);
+  };
+
   /**
    Convert to the ember-data REST adapter specification
